% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stability_plot.R
\name{create_grid}
\alias{create_grid}
\title{Create control grid for \code{starbility} estimation.}
\usage{
create_grid(perm, lhs, rhs, ...)
}
\arguments{
\item{perm}{A named dictionary in which values correspond to the sets of variables
that should be iterated upon to produce the stability plot and names correspond to the names
of these sets of variables that should be displayed in the plot.}

\item{lhs}{A string indicating the name of the outcome variable in \code{data}.}

\item{rhs}{A string indicating the name of the explanatory variable for which coefficient estimates
will be plotted.}

\item{base}{Optional. A named dictionary in which values correspond to the sets of variables
that should always be included in the model in all specifications and names correspond to the names
of these sets of variables that should be displayed in the plot.}

\item{perm_fe}{Optional. A named dictionary in which values correspond to the sets of fixed effects
that should be iterated upon to produce the stability plot and names correspond to the names
of these sets of variables that should be displayed in the plot. Functionally, these operate
identically to \code{perm}; the difference is that \code{starbility} uses \code{lfe} to sweep
them out of the normal equations, resulting in a performance boost over including them in
\code{perm}.}

\item{nonperm_fe}{Optional. A named dictionary in which values correspond to fixed effects that should be
iterated upon to produce the stability plot and names correspond to the names of these
sets of fixed effects that should be displayed in the plot. These fixed effects are included
sequentially in the plot, one at a time -- i.e. combinations of \code{nonperm_fe} are not included.}

\item{iv}{Optional. A string indicating the variables which should be used to instrument \code{rhs}.
If left unspecified, OLS coefficients are plotted.}

\item{fe_always}{Optional. A logical scalar. If one or more sets of fixed effects are
specified in \code{nonperm_fe}, should the plot include only estimates from models with
non-permuted fixed effects (rather than also including a set of estimates from models without
any non-permuted fixed effects)? Defaults to \code{F}.}
}
\value{
A data.frame containing all possible permutations of controls.
}
\description{
\code{create_grid} is the first step of a \code{starbility} plot. It is generally called
directly by \code{stability_plot}, but the user can also call it manually. This is generally
done when creating custom formulas that work with packages other than those automatically
supported by \\code{starbility} (currently, \code{starbility} supports only \code{felm}).
}
