% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stability_plot.R
\name{create_model_estimates}
\alias{create_model_estimates}
\title{Estimate control grid models.}
\usage{
create_model_estimates(grid., data., lhs., rhs., perm., ...)
}
\arguments{
\item{...}{Optional. Additional parameters to be passed to \code{model}.}

\item{grid}{A dataframe generated by \code{create_felm_formulas}, or a data frame of the same structure
as those created by \code{create_felm_formulas}.}

\item{data}{A dataframe containing the variables included in the model.}

\item{lhs}{A string indicating the name of the outcome variable in \code{data}.}

\item{rhs}{A string indicating the name of the explanatory variable for which coefficient estimates
will be plotted.}

\item{perm}{A named dictionary in which values correspond to the sets of variables
that should be iterated upon to produce the stability plot and names correspond to the names
of these sets of variables that should be displayed in the plot.}

\item{model}{Optional. A function that takes at least three arguments: \code{spec} (a string
containing the model specification), \code{data} (the data frame containing the variables in
the model), and \code{rhs} (the name of the coefficient of interest). Arbitrary additional
arguments are permitted. The function should then output a vector containing, in order, the
coefficient estimate, the p-value, the bottom value of the error region, and the top value of
the error region. If left unspecified, uses default implementation of \code{felm} (from \code{lfe}).}
}
\value{
A data.frame containing model estimates.
}
\description{
\code{create_model_estimates} is the third step of a \code{starbility} plot, following
\code{create_felm_formulas}. It is generally called directly by \code{stability_plot},
but the user can also call it manually.
}
