% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stability_plot.R
\name{create_plot_dfs}
\alias{create_plot_dfs}
\title{Create data frames for \code{stabilityplot} plotting.}
\usage{
create_plot_dfs(grid, perm., ...)
}
\arguments{
\item{grid}{A dataframe generated by \code{create_model_estimates}, or a data frame of the same structure
as those created by \code{create_model_estimates}.}

\item{perm}{A named dictionary in which values correspond to the sets of variables
that should be iterated upon to produce the stability plot and names correspond to the names
of these sets of variables that should be displayed in the plot.}

\item{base}{Optional. A named dictionary in which values correspond to the sets of variables
that should always be included in the model in all specifications and names correspond to the names
of these sets of variables that should be displayed in the plot.}

\item{perm_fe}{Optional. A named dictionary in which values correspond to the sets of fixed effects
that should be iterated upon to produce the stability plot and names correspond to the names
of these sets of variables that should be displayed in the plot. Functionally, these operate
identically to \code{perm}; the difference is that \code{starbility} uses \code{lfe} to sweep
them out of the normal equations, resulting in a performance boost over including them in
\code{perm}.}

\item{nonperm_fe}{Optional. A named dictionary in which values correspond to fixed effects that
should be iterated upon to produce the stability plot and names correspond to the names of these
sets of fixed effects that should be displayed in the plot. These fixed effects are included
sequentially in the plot, one at a time -- i.e. combinations of \code{nonperm_fe} are not
included.}

\item{sort}{A string specifying how models should be sorted by coefficient value. The default is
\code{none}, which preserves the order in which controls are permuted. Other options are
\code{asc} (sorted by ascending coefficient values), \code{desc} (sorted by descending
coefficient values), \code{asc-by-fe} (sorted by ascending coefficient values within non-permuted
fixed effects groups, but preserving the order of these groups), and \code{desc-by-fe}
(sorted by descending coefficient values within non-permuted fixed effects groups, but preserving the
order of these groups).}
}
\value{
A list containing two data.frames: one corresponding to the coefficient panel and one
corresponding to the control panel.
}
\description{
\code{create_plot_dfs} is the fourth step of a \code{starbility} plot, following
\code{create_model_estimates}. It is generally called directly by \code{stability_plot},
but the user can also call it manually.
}
