% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stability_plot.R
\name{create_stability_plot}
\alias{create_stability_plot}
\title{Create coefficient stability plot.}
\usage{
create_stability_plot(data, lhs, rhs, perm, ...)
}
\arguments{
\item{data}{A dataframe containing the variables in the model will be estimated.}

\item{lhs}{A string indicating the name of the outcome variable in \code{data}.}

\item{rhs}{A string indicating the name of the explanatory variable for which coefficient estimates
will be plotted.}

\item{perm}{A named dictionary in which values correspond to the sets of variables
that should be iterated upon to produce the stability plot and names correspond to the names
of these sets of variables that should be displayed in the plot.}

\item{iv}{A string indicating the variables which should be used to instrument \code{rhs}.
If left unspecified, OLS coefficients are plotted.}

\item{base}{A named dictionary in which values correspond to the sets of variables that should
always be included in the model in all specifications and names correspond to the names
of these sets of variables that should be displayed in the plot.}

\item{perm_fe}{A named dictionary in which values correspond to the sets of fixed effects
that should be iterated upon to produce the stability plot and names correspond to the names
of these sets of variables that should be displayed in the plot. Functionally, these operate
identically to \code{perm}; the difference is that \code{starbility} uses \code{lfe} to sweep
them out of the normal equations, resulting in a performance boost over including them in
\code{perm}.}

\item{nonperm_fe}{A named dictionary in which values correspond to fixed effects that should be
iterated upon to produce the stability plot and names correspond to the names of these
sets of fixed effects that should be displayed in the plot. These fixed effects are included
sequentially in the plot, one at a time -- i.e. combinations of \code{nonperm_fe} are not included.}

\item{include_no_fe_specs}{A logical scalar. If one or more sets of fixed effects are
specified in \code{nonperm_fe}, should the plot additionally include estimates from models without
non-permuted fixed effects? Defaults to \code{T}.}

\item{sort}{A string specifying how models should be sorted by coefficient value. The default is
\code{none}, which preserves the order in which controls are permuted. Other options are
\code{asc} (sorted by ascending coefficient values), \code{desc} (sorted by descending
coefficient values), \code{asc-by-fe} (sorted by ascending coefficient values within non-permuted
fixed effects groups, but preserving the order of these groups), and \code{desc-by-fe}
(sorted by descending coefficient values within non-permuted fixed effects groups, but preserving the
order of these groups).}

\item{cluster}{A string indicating the name of the variable by which standard errors should be
clustered. Defaults to no clustering.}

\item{weights}{A string indicating the name of the variable containing weights. Defaults to equal
weighting.}

\item{point_size}{A numeric scalar indicating the size of the points indicating coefficient estimates.
Defaults to 1.}

\item{error_geom}{A string indicating the type of geom that should be used to indicate confidence
intervals on coefficient estimates. Currently supported are \code{ribbon}, \code{errorbar}, and \code{none}.
Defaults to \code{errorbar} if fewer than 100 models are plotted; defaults to \code{ribbon} if
100 or more models are plotted.}

\item{error_alpha}{A numeric scalar indicating the alpha of the error geom. Defaults to 0.2.}

\item{coef_ylim}{A numeric vector of length two indicating the minimum and maximum values of the
y-axis in the coefficient plot. If not specified, uses \code{ggplot2} default.}

\item{control_geom}{A string indicating the geom that should be used to indicate the presence of
controls. Currently supported are \code{circle} and \code{rect}. Defaults to \code{rect}.}

\item{control_spacing}{A string indicating how large the geoms indicating the presence of controls
should be. For \code{control_geom=='circle'}, this is the diameter of the circle. For
\code{control_geom=='rect'}, this is the width of the rectangle. Defaults to 0.75 if fewer than
40 models are displayed; defaults to 1 otherwise.}

\item{control_text_size}{A numeric scalar indicating how large the control name text
should be. Defaults to 9.}

\item{rel_height}{A numeric scalar indicating the size of the bottom panel (displaying presence of
controls) relative to the top panel (displaying presence of coefficients). Defaults to 0.25.}

\item{trim_top}{A numeric scalar indicating how close the bottom panel (displaying presence of
controls) should be to the top panel (displaying presence of coefficients). Useful when dealing with
large CIs.}

\item{combine}{A logical scalar. Return the panels combined as a single object, or return a list
containing the two panels separately?}
}
\value{
A ggplot2 object.
}
\description{
\code{create_stability_plot} is used to produce a plot showing the stability of the OLS estimate
of the explanatory variable \code{rhs} on the outcome variable \code{lhs} under combinations
of a given set of controls. Fixed effects, clustering, weights, and instrumental
variables are supported.
}
\details{
Each row of the bottom panel of the plot corresponds to a single variable set.
A variable set can contain one or more individual variables.
To include multiple variables in a single set, specify them in a single string, separated by '+'.
}
