% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stability_plot.R
\name{draw_panels}
\alias{draw_panels}
\title{Draw coefficient stability plots.}
\usage{
draw_panels(coef_grid., control_grid., ...)
}
\arguments{
\item{coef_grid}{A dataframe generated by \code{create_grid} (first element of returned list),
or a data frame of the same structure.}

\item{control_grid}{A dataframe generated by \code{create_grid} (second element of returned list),
or a data frame of the same structure.}

\item{point_size}{A numeric scalar indicating the size of the points indicating coefficient estimates.
Defaults to 1.}

\item{error_geom}{A string indicating the type of geom that should be used to indicate confidence
intervals on coefficient estimates. Currently supported are \code{ribbon}, \code{errorbar}, and \code{none}.
Defaults to \code{errorbar} if fewer than 100 models are plotted; defaults to \code{ribbon} if
100 or more models are plotted.}

\item{error_alpha}{A numeric scalar indicating the alpha of the error geom. Defaults to 0.2.}

\item{coef_ylim}{A numeric vector of length two indicating the minimum and maximum values of the
y-axis in the coefficient plot. If not specified, uses \code{ggplot2} default.}

\item{coef_ylabel}{A string specifying the y-axis label on the coefficient panel. Defaults to
'Coefficient estimate'.}

\item{control_geom}{A string indicating the geom that should be used to indicate the presence of
controls. Currently supported are \code{circle} and \code{rect}. Defaults to \code{rect}.}

\item{control_spacing}{A string indicating how large the geoms indicating the presence of controls
should be. For \code{control_geom=='circle'}, this is the diameter of the circle. For
\code{control_geom=='rect'}, this is the width of the rectangle. Defaults to 0.75 if fewer than
40 models are displayed; defaults to 1 otherwise.}

\item{control_text_size}{A numeric scalar indicating how large the control name text
should be. Defaults to 9.}

\item{rel_height}{A numeric scalar indicating the size of the bottom panel (displaying presence of
controls) relative to the top panel (displaying presence of coefficients). Defaults to 0.25.}

\item{trim_top}{A numeric scalar indicating how close the bottom panel (displaying presence of
controls) should be to the top panel (displaying presence of coefficients). Useful when dealing with
large CIs.}

\item{combine}{A logical scalar. Return the panels combined as a single object, or return a list
containing the two panels separately?}
}
\value{
If \code{combine=T} (default) a \code{cowplot} grid with both panels. If \code{combine=F}, a list
containing two `ggplot2`` objects: one corresponding to the coefficient panel and one
corresponding to the control panel.
}
\description{
\code{draw_panels} is the fifth and final step of a \code{starbility} plot, following
\code{create_plot_dfs}. It is generally called directly by \code{stability_plot},
but the user can also call it manually.
}
